admin: Add Chart Catalog infrastructure
To support creating a catalog of predefined charts to aid in debugging, this PR introduces the
internal mechanisms by which we can define and generate a catalog of charts.

This PR does not contain an exhaustive catalog, but instead only an illustrative subset of the
catalog to demonstrate the code's ergonomics. I will add the remainder of the catalog once
this commit is reviewed.

The provided tests will also be expanded on in a subsequent PR.

Release note: none